+++ RELEASE-NOTES.md
@@ -46,7 +46,9 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
   delimiter characters
 - Inlet: Added a method to `inlet::Inlet` that retrieves the set of unexpected names
   in the input file
-- Added support for registering material and species sets in `MFEMSidreDataCollection`
+- Added support for registering material and species sets in `MFEMSidreDataCollection`.
+  These correspond to [`matset`](https://llnl-conduit.readthedocs.io/en/latest/blueprint_mesh.html#material-sets)s
+  and [`specset`](https://llnl-conduit.readthedocs.io/en/latest/blueprint_mesh.html#species-sets)s in the Mesh Blueprint
 
 ### Changed
 - The Sidre Datastore no longer rewires Conduit's error handlers to SLIC by default.